feat(android): anchor the profile menu to its button
The sheet slid up from the bottom of the screen, far from the avatar
that opened it, and gave two actions the full width of the display. A
menu hanging off the button keeps the control and what it opens
together, the way the Apple clients present the same thing.

The name still frames the actions, now as a heading above them instead
of a row of its own.
