--- kotlin/android/app/src/main/java/dev/firezone/android/features/session/ui/compose/ProfileSheet.kt
@@ -1,95 +1,0 @@
-// Licensed under Apache 2.0 (C) 2026 Firezone, Inc.
-package dev.firezone.android.features.session.ui.compose
-
-import androidx.annotation.DrawableRes
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ListItem
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.unit.dp
-import dev.firezone.android.R
-
-@Composable
-fun ProfileButton(
-    actorName: String,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val label = stringResource(R.string.profile)
-
-    IconButton(onClick = onClick, modifier = modifier.semantics { contentDescription = label }) {
-        Box(
-            Modifier.size(32.dp).background(MaterialTheme.colorScheme.primary, CircleShape),
-            contentAlignment = Alignment.Center,
-        ) {
-            Text(
-                text = actorName.take(1).uppercase(),
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.onPrimary,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ProfileSheet(
-    actorName: String,
-    onSettings: () -> Unit,
-    onSignOut: () -> Unit,
-    onDismiss: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val sheetState = rememberModalBottomSheetState()
-
-    ModalBottomSheet(onDismissRequest = onDismiss, sheetState = sheetState, modifier = modifier) {
-        Column(Modifier.fillMaxWidth().padding(bottom = 16.dp)) {
-            Text(
-                text = actorName,
-                style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-            )
-
-            ProfileAction(icon = R.drawable.rounded_settings_black_24dp, label = stringResource(R.string.settings)) {
-                onDismiss()
-                onSettings()
-            }
-            ProfileAction(icon = R.drawable.rounded_logout_24dp, label = stringResource(R.string.sign_out)) {
-                onDismiss()
-                onSignOut()
-            }
-        }
-    }
-}
-
-@Composable
-private fun ProfileAction(
-    @DrawableRes icon: Int,
-    label: String,
-    onClick: () -> Unit,
-) {
-    ListItem(
-        headlineContent = { Text(label) },
-        modifier = Modifier.clickable(onClick = onClick),
-        leadingContent = { Icon(painterResource(icon), contentDescription = null) },
-    )
-}
