+++ kotlin/android/app/src/main/java/dev/firezone/android/features/session/ui/compose/SessionScreen.kt
@@ -86,12 +86,15 @@ fun SessionScreen(
             (selection as? Selection.Device)?.let { sel -> connectedDevices.firstOrNull { it.id == sel.id } }
         }
 
-    var profileOpen by rememberSaveable { mutableStateOf(false) }
     val profileName = actorName ?: stringResource(R.string.signed_in)
 
     Scaffold(
         modifier = modifier,
-        topBar = { FirezoneTopBar { ProfileButton(actorName = profileName, onClick = { profileOpen = true }) } },
+        topBar = {
+            FirezoneTopBar {
+                ProfileMenu(actorName = profileName, onSettings = onSettings, onSignOut = onSignOut)
+            }
+        },
     ) { innerPadding ->
         Column(Modifier.fillMaxSize().padding(innerPadding).padding(16.dp)) {
             // The tab bar is the top-level switcher, pinned below the app bar so it stays visible and
@@ -162,15 +165,6 @@ fun SessionScreen(
         ConnectedDeviceDetailsSheet(
             device = device,
             onDismiss = { selection = null },
-        )
-    }
-
-    if (profileOpen) {
-        ProfileSheet(
-            actorName = profileName,
-            onSettings = onSettings,
-            onSignOut = onSignOut,
-            onDismiss = { profileOpen = false },
         )
     }
 }
